+++ plugins/Makefile
@@ -37,6 +37,7 @@ PLUGINS :=					\
 
 PLUGIN_COMMON_OBJS :=				\
 	bitcoin/base58.o			\
+	bitcoin/feerate.o			\
 	bitcoin/privkey.o			\
 	bitcoin/psbt.o				\
 	bitcoin/pubkey.o			\
